Fix bug regarding shiny items
This PR aims to fix a bug where if a user attempts to shiny an item using the upgrade station and an offering, if it succeeds, the server does not send a response. This appears to occur because the server sets a "announce" flag to be false when the upgrade is a shiny attempt, and this prevents the success response from being sent to the client.