Fix Piercing Shot to actually grant Armor Piercing
Prior to this change, Piercing Shot (A ranger skill) that gives 500 armor piercing would only grant this pierce if the player had no pierce themselves. If the player had pierce, the pierce from the skill would then be overwritten by the player's pierce. This change fixes it by adding the player's pierce (or miss chance!) to the skills stats.