Fix another bug with Giveaways
If a player was not on the server when the giveaway ended, or their inventory was full, then the server would push `{ name: undefined }` into the list of players that could win the giveaway. This would send the item into the abyss, as it would error while the server would try to mail the item, preventing it from being recovered.